+++ docs/changelog.rst
@@ -1,6 +1,14 @@
 Changelog
 --------------
 
+**[0.3.0] - 2019-05-30**
+Added
+- Add `exact_order` param to `has_columns`
+
+Changed
+- Hotfix for reversed `has_columns` error messages for missing and unexpected columns
+- Breaking change to `has_columns` parameter name `exact`, which is now `exact_cols`
+
 **[0.2.0] - 2019-05-29**
 Added
 - Add `has_columns` check, which asserts that the given columns are contained within the df or exactly match the df's columns.